日常提交

+++ mayfly-dao/src/main/java/mayfly/entity/Permission.java
@@ -26,11 +26,11 @@ public class Permission implements Serializable {
 
     private String uriPattern;
 
-    private Byte method;
+    private Integer method;
 
     private Integer groupId;
 
-    private Byte status;
+    private Integer status;
 
     public LocalDateTime createTime;
